stable-coin-perf-test/requirements.txt
locust==1.2.3
grpcio==1.32.0
grpcio-tools==1.32.0
pynacl==1.3.0
